Do not depend on MPI imports
